requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
matplotlib>=3.7.0
folium>=0.14.0
pillow>=10.0.0
selenium>=4.15.0
numpy>=1.24.0